requirements.txt
crewai[google-genai]
PyPDF2
python-dotenv
streamlit
markdown2
